refactor(results.component.ts): two separate components
