+++ server/package.json
@@ -17,7 +17,7 @@
     "@types/mongoose": "^5.0.15",
     "@types/node": "^10.3.0",
     "nodemon": "^1.17.5",
-    "ts-node": "^6.1.0",
+    "ts-node": "^6.2.0",
     "typescript": "^2.9.1"
   },
   "dependencies": {
